wip: restart language server on files change
