+++ packages/vscode/src/index.ts
@@ -1,10 +1,5 @@
 import { effectScope, shallowRef } from '@vue/runtime-core'
 import type { ExtensionContext } from 'vscode'
-import { Uri } from 'vscode'
-import type { LanguageClientOptions, ServerOptions } from '@volar/vscode/node'
-import { LanguageClient, TransportKind } from '@volar/vscode/node'
-import { createLabsInfo } from '@volar/vscode'
-import * as serverProtocol from '../language-server/protocol'
 import { useCommands } from './commands'
 import { useGlobalConfigurations } from './configs'
 import { activeEntry, useProjects } from './projects'
@@ -14,17 +9,16 @@ import { useLogger } from './views/logger'
 import { usePreviewWebview } from './views/previewWebview'
 import { useSlidesTree } from './views/slidesTree'
 import { useProjectsTree } from './views/projectsTree'
+import { useLanguageClient } from './languageClient'
 
 const scope = effectScope()
-
-let client: LanguageClient
 
 export const extCtx = shallowRef<ExtensionContext>(undefined!)
 
 export async function activate(ext: ExtensionContext) {
   extCtx.value = ext
 
-  scope.run(() => {
+  return scope.run(() => {
     // states
     useGlobalConfigurations()
     useProjects()
@@ -39,44 +33,15 @@ export async function activate(ext: ExtensionContext) {
     useAnnotations()
     useFoldings()
 
+    // language server
+    const labsInfo = useLanguageClient()
+
     const logger = useLogger()
     logger.info('Slidev activated.')
     logger.info(`Entry: ${activeEntry.value}`)
+
+    return labsInfo
   })
-
-  return await startLanguageServer(ext)
-}
-
-async function startLanguageServer(ext: ExtensionContext) {
-  const serverModule = Uri.joinPath(ext.extensionUri, 'dist', 'language-server.cjs')
-  const runOptions = { execArgv: <string[]>[] }
-  const debugOptions = { execArgv: ['--nolazy', `--inspect=${6009}`] }
-  const serverOptions: ServerOptions = {
-    run: {
-      module: serverModule.fsPath,
-      transport: TransportKind.ipc,
-      options: runOptions,
-    },
-    debug: {
-      module: serverModule.fsPath,
-      transport: TransportKind.ipc,
-      options: debugOptions,
-    },
-  }
-  const clientOptions: LanguageClientOptions = {
-    documentSelector: [{ language: 'markdown' }],
-  }
-  client = new LanguageClient(
-    'slidev-language-server',
-    'Slidev Language Server',
-    serverOptions,
-    clientOptions,
-  )
-  await client.start()
-
-  const labsInfo = createLabsInfo(serverProtocol)
-  labsInfo.addLanguageClient(client)
-  return labsInfo.extensionExports
 }
 
 export async function deactivate() {
